fix(cli): close design proposal verification regressions
