+++ orchestration/schemas/code-intel-design-context.v1.schema.json
@@ -1,0 +1,42 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "code-intel-design-context.v1",
+  "title": "Code Intel evidence-bound design context",
+  "type": "object",
+  "additionalProperties": false,
+  "required": ["schema", "type", "snapshot", "evidenceRefs", "methods", "constraints", "knownUnknowns"],
+  "properties": {
+    "schema": { "const": "code-intel-design-context.v1" },
+    "type": { "const": "design.context" },
+    "snapshot": { "$ref": "#/$defs/snapshot" },
+    "evidenceRefs": { "$ref": "#/$defs/evidenceRefs" },
+    "methods": { "$ref": "#/$defs/stringArray" },
+    "constraints": { "$ref": "#/$defs/stringArray" },
+    "knownUnknowns": { "$ref": "#/$defs/stringArray" }
+  },
+  "$defs": {
+    "digest": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
+    "repoIdentity": { "type": "string", "pattern": "^(git-lineage-v1|content-v1):[0-9a-f]{64}$" },
+    "evidenceRefs": {
+      "type": "array",
+      "items": { "type": "string", "pattern": "^artifact://sha256/[0-9a-f]{64}$" }
+    },
+    "stringArray": {
+      "type": "array",
+      "items": { "type": "string" }
+    },
+    "snapshot": {
+      "type": "object",
+      "additionalProperties": false,
+      "required": ["identity", "repoIdentity", "head", "workingTreePolicy", "scope", "inputDigest"],
+      "properties": {
+        "identity": { "$ref": "#/$defs/digest" },
+        "repoIdentity": { "$ref": "#/$defs/repoIdentity" },
+        "head": { "type": "string", "minLength": 1 },
+        "workingTreePolicy": { "enum": ["head_only", "explicit_overlay"] },
+        "scope": { "type": "array", "minItems": 1, "items": { "type": "string", "minLength": 1 } },
+        "inputDigest": { "$ref": "#/$defs/digest" }
+      }
+    }
+  }
+}
