+++ orchestration/schemas/code-intel-design-proposal-candidate.v1.schema.json
@@ -1,0 +1,96 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "code-intel-design-proposal-candidate.v1",
+  "title": "Code Intel evidence-bound design proposal candidate",
+  "type": "object",
+  "additionalProperties": false,
+  "required": ["schema", "kind", "authority", "snapshot", "request", "baseline", "delta", "methods", "options", "recommendation", "risks", "validationPlan", "limitations"],
+  "properties": {
+    "schema": { "const": "code-intel-design-proposal-candidate.v1" },
+    "kind": { "const": "design_proposal_candidate" },
+    "authority": { "const": "advisory_only" },
+    "snapshot": { "$ref": "#/$defs/snapshot" },
+    "request": { "$ref": "#/$defs/request" },
+    "baseline": { "$ref": "#/$defs/section" },
+    "delta": { "$ref": "#/$defs/section" },
+    "methods": {
+      "type": "array",
+      "items": {
+        "type": "object",
+        "additionalProperties": false,
+        "required": ["id", "evidenceRefs"],
+        "properties": {
+          "id": { "type": "string", "minLength": 1 },
+          "evidenceIds": { "$ref": "#/$defs/nonemptyStrings" },
+          "evidenceRefs": { "$ref": "#/$defs/evidenceRefs" }
+        }
+      }
+    },
+    "options": { "type": "array", "minItems": 2, "maxItems": 3, "items": { "$ref": "#/$defs/option" } },
+    "recommendation": { "$ref": "#/$defs/recommendation" },
+    "risks": { "$ref": "#/$defs/stringArray" },
+    "validationPlan": { "$ref": "#/$defs/stringArray" },
+    "limitations": { "$ref": "#/$defs/stringArray" }
+  },
+  "$defs": {
+    "digest": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
+    "repoIdentity": { "type": "string", "pattern": "^(git-lineage-v1|content-v1):[0-9a-f]{64}$" },
+    "evidenceRefs": {
+      "type": "array", "minItems": 1,
+      "items": { "type": "string", "pattern": "^artifact://sha256/[0-9a-f]{64}$" }
+    },
+    "nonemptyStrings": { "type": "array", "minItems": 1, "items": { "type": "string", "minLength": 1 } },
+    "stringArray": { "type": "array", "items": { "type": "string" } },
+    "snapshot": {
+      "type": "object", "additionalProperties": false,
+      "required": ["identity", "repoIdentity", "head", "workingTreePolicy", "scope", "inputDigest"],
+      "properties": {
+        "identity": { "$ref": "#/$defs/digest" },
+        "repoIdentity": { "$ref": "#/$defs/repoIdentity" },
+        "head": { "type": "string", "minLength": 1 },
+        "workingTreePolicy": { "enum": ["head_only", "explicit_overlay"] },
+        "scope": { "type": "array", "minItems": 1, "items": { "type": "string", "minLength": 1 } },
+        "inputDigest": { "$ref": "#/$defs/digest" }
+      }
+    },
+    "request": {
+      "type": "object", "additionalProperties": false,
+      "required": ["mode", "capability", "schema"],
+      "properties": {
+        "mode": { "const": "validate" },
+        "capability": { "const": "advisory.design-proposal.compat" },
+        "schema": { "const": "code-intel-design-proposal-request.v1" }
+      }
+    },
+    "section": {
+      "type": "object", "additionalProperties": false,
+      "required": ["summary", "evidenceRefs"],
+      "properties": {
+        "summary": { "type": "string", "minLength": 1 },
+        "evidenceRefs": { "$ref": "#/$defs/evidenceRefs" }
+      }
+    },
+    "option": {
+      "type": "object", "additionalProperties": false,
+      "required": ["id", "title", "summary", "boundaryChanges", "tradeoffs", "assumptions", "evidenceRefs", "validationPlan", "reversibility"],
+      "properties": {
+        "id": { "type": "string", "minLength": 1 },
+        "title": { "type": "string", "minLength": 1 },
+        "summary": { "type": "string", "minLength": 1 },
+        "boundaryChanges": { "$ref": "#/$defs/stringArray" },
+        "tradeoffs": { "$ref": "#/$defs/stringArray" },
+        "assumptions": { "$ref": "#/$defs/stringArray" },
+        "evidenceRefs": { "$ref": "#/$defs/evidenceRefs" },
+        "validationPlan": { "$ref": "#/$defs/stringArray" },
+        "reversibility": {
+          "type": "object", "additionalProperties": false, "required": ["status", "basis"],
+          "properties": { "status": { "type": "string", "minLength": 1 }, "basis": { "type": "string", "minLength": 1 } }
+        }
+      }
+    },
+    "recommendation": {
+      "type": "object", "additionalProperties": false, "required": ["optionId", "rationale"],
+      "properties": { "optionId": { "type": "string", "minLength": 1 }, "rationale": { "type": "string", "minLength": 1 } }
+    }
+  }
+}
